refactor: define phase P ownership boundaries
